requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==23.1.0
bleach==5.0.1
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmake==3.25.0
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastai==2.7.10
fastcore==1.5.27
fastdownload==0.0.7
fastjsonschema==2.16.2
fastprogress==1.0.3
filelock==3.9.0
fonttools==4.38.0
fvcore==0.1.5.post20221221
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
hydra-core==1.3.2
idna==3.4
ijson==3.1.4
imageio==2.22.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iopath==0.1.9
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-client==7.4.7
jupyter-console==6.4.4
jupyter-core==5.0.0
jupyter-server==1.23.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kaggle==1.5.12
kiwisolver==1.4.4
langcodes==3.3.0
lit==15.0.7
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.2.1
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pathy==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.4
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==4.22.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-slugify==7.0.0
pytorch-triton==2.0.0+b8b470bc59
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
stack-data==0.6.1
swig==4.1.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.0
text-unidecode==1.3
thinc==8.1.5
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typer==0.7.0
typing-extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
widgetsnbextension==4.0.3
yacs==0.1.8
zipp==3.10.0
